refactor: course details hero section
