+++ src/components/CourseDetailsComponent/CourseHeroSection.js
@@ -1,32 +1,10 @@
-import { makeStyles, Box, Typography, Container, Button } from "@material-ui/core";
+import { makeStyles, Box, Typography, Container } from "@material-ui/core";
 import React from "react";
 import NavBar from "../../components/NavBar/index";
-import KeyboardArrowRightIcon from "@material-ui/icons/KeyboardArrowRight";
 import Rating from "@material-ui/lab/Rating";
-import ReactPlayer from "react-player/lazy";
-
-const pathList = [
-  {
-    text: " Development",
-    path: "",
-    isArrow: true,
-  },
-  {
-    text: " Web Development",
-    path: "",
-    isArrow: true,
-  },
-  {
-    text: "Training",
-    path: "",
-    isArrow: true,
-  },
-  {
-    text: " React",
-    path: "",
-    isArrow: false,
-  },
-];
+import VideoCard from "./SubComponents/HeroElements/VideoCard";
+import CoursePath from "./SubComponents/HeroElements/CoursePath";
+import ShareAndWishlistButton from "./SubComponents/HeroElements/ShareAndWishlistButton";
 function CourseHeroSection() {
   const classes = useStyles();
 
@@ -35,16 +13,7 @@ function CourseHeroSection() {
       <NavBar />
       <Box pt={14} className={classes.root}>
         <Container className={classes.innerContainer}>
-          <Box className={classes.pathContainer}>
-            {pathList.map((items, index) => (
-              <>
-                <Typography variant="h6" key={index} className={classes.pathText}>
-                  {items.text}
-                </Typography>
-                {items.isArrow ? <KeyboardArrowRightIcon /> : <Box />}
-              </>
-            ))}
-          </Box>
+          <CoursePath /> {/* course path component */}
           <Box mt={4}>
             <Typography gutterBottom className={classes.title}>
               Machine Learning, Data Science and Deep Learning with Python
@@ -60,24 +29,9 @@ function CourseHeroSection() {
               <Typography className={classes.numberOfStudents}> 148,806 students </Typography>
             </Box>
           </Box>
-          <Box mt={4} className={classes.buttonContainer}>
-            {[1, 2].map((items, index) => (
-              <Button key={index} className={classes.button}>
-                <Typography> Share</Typography>
-              </Button>
-            ))}
-          </Box>
+          <ShareAndWishlistButton /> {/* Share and add wishlist button component */}
         </Container>
-        <Container className={classes.videoContainer}>
-          <ReactPlayer
-            className={classes.reactPlayer}
-            controls={true}
-            light="https://images.unsplash.com/photo-1423530990185-af7dde1155fb?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=2000&q=80"
-            url="https://test-videos.co.uk/vids/bigbuckbunny/mp4/h264/360/Big_Buck_Bunny_360_10s_1MB.mp4"
-            volume={1}
-            width="100%"
-          />
-        </Container>
+        <VideoCard /> {/* video card component */}
       </Box>
     </>
   );
@@ -88,18 +42,8 @@ const useStyles = makeStyles((theme) => ({
     height: "585px",
     color: theme.palette.text.main,
     display: "flex",
-    paddingLeft: theme.spacing(5),
-    paddingRight: theme.spacing(5),
-  },
-  pathContainer: {
-    display: "flex",
-    color: theme.palette.primary.main,
-  },
-  pathText: {
-    "&:hover": {
-      textDecoration: "underline",
-      cursor: "pointer",
-    },
+    paddingLeft: theme.spacing(10),
+    paddingRight: theme.spacing(10),
   },
   ratingContainer: {
     display: "flex",
@@ -118,21 +62,6 @@ const useStyles = makeStyles((theme) => ({
   title: {
     fontSize: "2rem",
     fontWeight: 600,
-  },
-  button: {
-    color: "#fff",
-    textTransform: "none",
-    border: "2px solid #fff",
-    marginRight: theme.spacing(2),
-    padding: theme.spacing(1, 4),
-  },
-  videoContainer: {
-    flex: 2,
-    // background: "#000",
-    marginLeft: theme.spacing(10),
-    marginRight: theme.spacing(5),
-    marginBottom: theme.spacing(5),
-    borderRadius: "5px",
   },
 }));
 export default CourseHeroSection;
