+++ src/components/CourseDetailsComponent/SubComponents/HeroElements/CoursePath.js
